+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v0.1.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds email verification
 - Adds admin user
 - Adds `create user` function to cli
 - Adds user module